pyproject.toml
[project]
name = "hackaton2025-counterpart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.66.3",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
]
